requirements.txt
"""
pandas==2.1.4
yfinance==0.2.28
python-dotenv==1.0.0
loguru==0.7.2
SQLAlchemy==2.0.25
requests==2.31.0
"""